+++ extensions/raycast-wallpaper/src/components/raycast-wallpaper-list.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import { RaycastWallpaperWithInfo } from "../types/types";
 import { RaycastWallpaperEmptyView } from "./raycast-wallpaper-empty-view";
 import { ActionOnRaycastWallpaper } from "./action-on-raycast-wallpaper";
-import { getThumbnailUrl } from "../utils/common-utils";
+import { capitalizeFirstLetter, getThumbnailUrl } from "../utils/common-utils";
 import { layout } from "../types/preferences";
 
 export function RaycastWallpaperList(props: {
@@ -28,6 +28,7 @@ export function RaycastWallpaperList(props: {
     >
       <RaycastWallpaperEmptyView layout={layout} />
       {raycastWallpapers.map((value, index) => {
+        const appearanceIcon = value.appearance == "light" ? Icon.Sun : Icon.Moon;
         return (
           <List.Item
             id={index + ""}
@@ -35,7 +36,9 @@ export function RaycastWallpaperList(props: {
             icon={{ source: getThumbnailUrl(value.url) }}
             title={value.title}
             accessories={
-              value.exclude ? [{ icon: Icon.XMarkTopRightSquare, tooltip: "Excluded From Auto Switch" }] : []
+              value.exclude
+                ? [{ icon: Icon.XMarkTopRightSquare, tooltip: "Excluded From Auto Switch" }]
+                : [{ icon: appearanceIcon, tooltip: capitalizeFirstLetter(value.appearance) }]
             }
             detail={<List.Item.Detail isLoading={false} markdown={`![](${getThumbnailUrl(value.url)})`} />}
             actions={