+++ extensions/raycast-wallpaper/src/types/preferences.ts
@@ -2,7 +2,8 @@ import { getPreferenceValues } from "@raycast/api";
 
 export const { layout, columns, picturesDirectory } = getPreferenceValues<Preferences.SetRaycastWallpaper>();
 
-export const { refreshIntervalSeconds } = getPreferenceValues<Preferences.AutoSwitchRaycastWallpaper>();
+export const { refreshIntervalSeconds, respectAppearance } =
+  getPreferenceValues<Preferences.AutoSwitchRaycastWallpaper>();
 
 export interface DuplicatePreferences {
   applyTo: string;